Update README.md
Added badges for NPM version & total downloads.

If you land on the Github page of openapi-to-postman, it is confusing to find out what is now the latest version at first glance.
Adding the NPM version badge will make it easier. The total downloads badge indicates how widely used the package is.